Refactor: Reducing Scope
Found out last night that I didn't have time to do everything I wanted before the end of the week. Removed the language select and removed the parts of the script that was meant for other languages and principle quiz as I only will focus on JS and its syntax.